bug creating book fixed
