+++ data/labels.json
@@ -1,1 +1,13 @@
-[{"id":38360,"title":"New","color":"Green","items":[{"id":42558,"class":"Book"}]}]
+[
+  {
+    "id": 61090,
+    "title": "New",
+    "color": "Grae",
+    "items": [
+      {
+        "id": 47301,
+        "class": "Book"
+      }
+    ]
+  }
+]
